move type to TableDataSourceConfiguration front/lib/api/assistant + more cleanup
