+++ front/lib/actions/types/guards.ts
@@ -15,7 +15,6 @@ import type {
 import { isInternalMCPServerOfName } from "@app/lib/actions/mcp_internal_actions/constants";
 import type { ProcessConfigurationType } from "@app/lib/actions/process";
 import type { SearchLabelsConfigurationType } from "@app/lib/actions/search_labels";
-import type { TablesQueryConfigurationType } from "@app/lib/actions/tables_query";
 import type {
   ActionConfigurationType,
   AgentActionConfigurationType,
@@ -30,17 +29,6 @@ import type {
   AgentConfigurationType,
   TemplateAgentConfigurationType,
 } from "@app/types";
-
-export function isTablesQueryConfiguration(
-  arg: unknown
-): arg is TablesQueryConfigurationType {
-  return (
-    !!arg &&
-    typeof arg === "object" &&
-    "type" in arg &&
-    arg.type === "tables_query_configuration"
-  );
-}
 
 export function isDustAppRunConfiguration(
   arg: unknown
